requirements.txt
boto3==1.26.108
schedule==1.2.0
requests==2.28.2
python-dotenv==1.0.0
